+++ tests/test_draw.py
@@ -9,7 +9,7 @@ from mmodel.model import Model
 import pytest
 
 dot_source = """digraph test {
-graph [label=test labelloc=t ordering=out splines=ortho]
+graph [label="test\ntest object" labelloc=t ordering=out splines=ortho]
 node [shape=box]
 add
 subtract
@@ -28,14 +28,14 @@ multiply -> poly
 def test_draw_model_no_detail(mmodel_graph):
     """Test the model without the node detail"""
 
-    dot_graph = draw_graph(mmodel_graph, title="test", show_detail=False)
+    dot_graph = draw_graph(mmodel_graph, name="test", show_detail=False)
     assert dot_graph.source.replace('\n', '').replace('\t', '') == dot_source.replace('\n', '')
 
 
 def test_draw_model_detail(mmodel_graph):
     """Test the model with full node and edge detail"""
 
-    dot_graph = draw_graph(mmodel_graph, title="test")
+    dot_graph = draw_graph(mmodel_graph, name="test")
     assert "add\l\naddition(a, b=2)\lreturn c\l" in dot_graph.source
 
 
@@ -44,7 +44,7 @@ def test_draw_model_settings(mmodel_graph):
 
     dot_graph = draw_graph(
         mmodel_graph,
-        title="test",
+        name="test",
         node_attr={"shape": "plaintext"},
         edge_attr={"label": "test"},
     )
@@ -58,12 +58,12 @@ def test_draw_model_settings(mmodel_graph):
 @pytest.fixture
 def model_instance(mmodel_graph):
     """Create Model object for the test"""
-    return Model(mmodel_graph, "model_test")
+    return Model(mmodel_graph)
 
 
 def test_draw_graph_subgraph(model_instance):
     """Test change model settings
-    
+ 
     It is hard to test the dot file because the order of
     the nodes is not garanteed.
     """
@@ -71,10 +71,10 @@ def test_draw_graph_subgraph(model_instance):
     model_instance.loop_parameter(params=["f"])
     dot_graph = draw_graph(
         model_instance.graph,
-        "subgraph_test",
         show_detail=False,
         filename="test.gv",
         format="svg",
     )
 
-    assert "subgraph cluster_f_loop_submodel" in dot_graph.source
+    assert dot_graph.name == "test"
+    assert "subgraph cluster_basic_loop_f" in dot_graph.source