+++ CHANGELOG.rst
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_
 
+[unreleased]
+--------------------
+Added
+^^^^^
+- ``doc`` attribute for ``MGraph``
+
+Changed
+^^^^^^^
+- remove ``name`` input for ``Model`` and ``loop_parameter``
+- generate model names and looped subgraph name automatically
+- remove ``title`` input for ``draw_graph``
+
 
 [0.1.0] - 2022-04-02
 --------------------